Login-Consumer/requirements.txt
Flask==2.1.3
requests==2.28.1
PyJWT==2.4.0